requirements.txt
WeasyPrint==56.1
Jinja2==3.1.2
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
slack-sdk==3.26.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
click==8.1.7
openai>=1.0.0
Flask==3.0.0
google-auth-oauthlib>=1.2.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
flake8>=7.3.0
autopep8>=2.0.4
